# docs/cinerating_openapi.yaml
openapi: 3.0.3
info:
  title: CineRating API
  version: 2.1.0
  description: API REST para gestao de avaliacoes de filmes. A autenticacao inicial usa OAuth 2.0 com GitHub e as rotas protegidas usam JWT Bearer Token emitido pela propria API.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token emitido pela API.
  parameters:
...
    WatchlistFilmeId:
      in: path
      name: id_filme
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    ErrorResponse:
      type: object
      required:
        - erro
      properties:
        erro:
          type: string
          example: Recurso nao encontrado.
        detalhes:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
    MessageResponse:
      type: object
      required:
        - mensagem
      properties:
        mensagem:
          type: string
          example: Operacao efetuada com sucesso.
...
    Genero:
      type: object
      properties:
        id_genero:
          type: integer
          example: 1
        nome_genero:
          type: string
          example: Ficcao Cientifica
...
    Filme:
      type: object
      properties:
        id_filme:
          type: integer
          example: 1
        titulo:
          type: string
          example: Oppenheimer
        ano_lancamento:
          type: integer
          example: 2023
        duracao_minutos:
          type: integer
          nullable: true
          example: 180
        sinopse:
          type: string
          nullable: true
          example: A historia do cientista que liderou o projecto da bomba atomica.
        id_genero:
          type: integer
          nullable: true
          example: 16
        genero:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Genero"
...
    Watchlist:
      type: object
      description: Item de watchlist identificado pela chave composta id_utilizador + id_filme.
      properties:
        id_utilizador:
          type: integer
          example: 1
        id_filme:
          type: integer
          example: 2
        data_adicionado:
          type: string
          format: date-time
          example: "2026-04-30T12:00:00.000Z"
        visto:
          type: boolean
          example: false
        filme:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Filme"
...
    CreateWatchlistRequest:
      type: object
      required:
        - id_filme
      properties:
        id_filme:
          type: integer
          minimum: 1
          example: 2
    UpdateWatchlistRequest:
      type: object
      required:
        - visto
      properties:
        visto:
          type: boolean
          example: true
...
  responses:
    BadRequest:
      description: Pedido invalido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Recurso nao encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflito com o estado actual dos dados.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /watchlist:
    get:
      tags:
        - Watchlist
      summary: Listar watchlist do utilizador autenticado
      description: O id_utilizador usado no filtro vem do JWT; o cliente nao consegue consultar watchlists de terceiros.
      responses:
        "200":
          description: Watchlist propria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watchlist"
    post:
      tags:
        - Watchlist
      summary: Adicionar filme a watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWatchlistRequest"
      responses:
        "201":
          description: Filme adicionado a watchlist.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  item:
                    $ref: "#/components/schemas/Watchlist"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /watchlist/{id_filme}:
    get:
      tags:
        - Watchlist
      summary: Obter item da propria watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistFilmeId"
      responses:
        "200":
          description: Item encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Watchlist
      summary: Actualizar item da propria watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistFilmeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWatchlistRequest"
      responses:
        "200":
          description: Item actualizado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  item:
                    $ref: "#/components/schemas/Watchlist"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags:
        - Watchlist
      summary: Actualizar parcialmente item da propria watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistFilmeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWatchlistRequest"
      responses:
        "200":
          description: Item actualizado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                  item:
                    $ref: "#/components/schemas/Watchlist"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Watchlist
      summary: Remover item da propria watchlist
      parameters:
        - $ref: "#/components/parameters/WatchlistFilmeId"
      responses:
        "200":
          description: Item removido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"